Fix/dynamic typing (#37)
* perform periodic storing of the yaml file.

This is useful to avoid data loss in case the node is killed by an
unexpected signal. By a SIGKILL for instance.

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>

rename frequency to period and review formatting

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>

implement reviewer comments

- throw if the given period is not strictly positive
- use unsigned int for the period
- don't make unecessary changes

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>

don't perform periodic persistent parameter storing if the period is 0

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>

* fix double parameter value cannot be set once loaded as an integer

When a parameter is loaded from yaml, its type cannot be changed latler
through a call to set_parameters serrvice. This is problematic for some
cases. The ParameterDescriptor's dynamic_typing is set to true to solve
this issue

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>

* allow the user to control the dynamic typing option

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>

* pass storing_period in options.parameter_overrides instead of class argument

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>

* refactor the tests and add new test series for dynamic typing

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>

* correctly report test result summary code

---------

Signed-off-by: Mahmoud Almasri <mahm.al.masri@gmail.com>