Added ros2 service call to manually trigger yaml save (#38)
* Add ROS2 Service to manually trigger save

Updates the parameter server's SaveYamlFile function to optionally check the content before saving, aborting if there is an error. The service call uses this to make sure the save was successful.

Signed-off-by: Andrew Hoener <ahoener@boston-engineering.com>

* Add ROS2 Service to trigger a reload from file

This is mostly used for testing, but could be useful if a user wants to discard parameter changes and reload.

Signed-off-by: Andrew Hoener <ahoener@boston-engineering.com>

* Add tests to make sure save & reload works properly

Signed-off-by: Andrew Hoener <ahoener@boston-engineering.com>

* Remove old/unused test.cpp as tests have been refactored

Signed-off-by: Andrew Hoener <ahoener@boston-engineering.com>

---------

Signed-off-by: Andrew Hoener <ahoener@boston-engineering.com>